Revert "V3.6.0 reflection refact (#12096)"
This reverts commit 5b9b98f759b1300440183b357bd3543d14ac4165.
