+++ platforms/native/engine/jsb-bridge.js
@@ -22,19 +22,22 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  THE SOFTWARE.
  ****************************************************************************/
-require('./jsb-reflection.js');
-require('./jsb-bridge.js');
-require('./jsb-bridge-wrapper.js');
-require('./jsb-assets-manager.js');
 
-require('./jsb-game.js');
-require('./jsb-gfx.js');
-require('./jsb-loader.js');
-require('./jsb-videoplayer.js');
-require('./jsb-webview.js');
-require('./jsb-editbox.js');
-require('./jsb-editor-support.js');
-require('./jsb-spine-skeleton.js');
-require('./jsb-dragonbones.js');
-
-if (cc.physics && cc.physics.PhysicsSystem.PHYSICS_PHYSX) { require('./jsb-physics.js'); }
+// JS to Native bridges
+// set to lazy
+Object.defineProperty(jsb, 'bridge', {
+    get () {
+        if (jsb.__ccbridge !== undefined) return jsb.__ccbridge;
+        if (window.ScriptNativeBridge && cc.sys.os === cc.sys.OS.ANDROID || cc.sys.os === cc.sys.OS.IOS || cc.sys.os === cc.sys.OS.OSX || cc.sys.os === cc.sys.OS.OHOS) {
+            jsb.__ccbridge = new ScriptNativeBridge();
+        } else {
+            jsb.__ccbridge = null;
+        }
+        return jsb.__ccbridge;
+    },
+    enumerable: true,
+    configurable: true,
+    set (value) {
+        jsb.__ccbridge = value;
+    },
+});
